adding basic cakephp test, fixing some test tools warnings (#263)
this adds some basic test coverage for CakePHP. I couldn't get cake's testing
framework running, as I couldn't find a phpunit version it is compatible with,
so I just created a controller and fumbled my way to running it something like
how cake would. We have some coverage now, and a todo for a future improvement.